Created Ruby on Rails example app.
This app uses both OmniAuth Azure AD and ADAL. It uses OmniAuth to authenticate the user. This authentication is used to allow users to manage their personal Todo Lists. Also, the app allows users to view their profile data. This makes an API call to the AzureAD Graph which requires an access token as authorization. The user model uses ADAL to acquire access tokens for the users that have been authenticated.